Split gen go files
Start making the code a little more manageable to navigate.
Add standard editor config - start moving toward prettier for standardised formatting.
